requirements.txt
amqp==5.2.0
asgiref==3.8.1
beautifulsoup4==4.12.3
billiard==4.2.1
boto3==1.35.44
botocore==1.35.44
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==43.0.1
dj-database-url==2.2.0
Django
django-allauth==65.0.2
django-bootstrap5==24.3
django-extensions==3.2.3
django-heroku==0.3.1
django-s3direct==2.0.3
django-storages==1.14.4
gunicorn==23.0.0
icalendar==4.0.9
idna==3.10
iniconfig==2.0.0
jmespath==1.0.1
kombu==5.4.2
oauthlib==3.2.2
packaging==24.1
Pillow==10.0.1
pluggy==1.5.0
prompt_toolkit==3.0.48
psycopg2==2.9.9
pycparser==2.22
PyJWT==2.9.0
pytest==8.3.3
python-dateutil==2.9.0.post0
pytz==2024.2
redis==5.2.0
requests==2.32.3
requests-oauthlib==2.0.0
s3transfer==0.10.3
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.7.0
